backend/requirements.txt
# Core Django
Django==5.1.7
djangorestframework==3.16.0
django-cors-headers==4.9.0
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Task Queue
redis==5.0.1
django-redis==5.4.0
hiredis==2.3.2
celery==5.3.6
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Production Server
gunicorn==21.2.0
whitenoise==6.7.0

# Google Cloud & Earth Engine
earthengine-api==1.6.6
google-api-python-client==2.182.0
google-auth==2.40.3
google-cloud-storage==3.4.0
google-generativeai==0.8.3

# Geospatial
shapely==2.0.6
pyproj==3.7.0

# ML & Data Science
torch==2.5.1
torchvision==0.20.1
numpy==2.1.2
pandas==2.2.3
scikit-learn==1.6.1
Pillow==11.3.0

# HTTP & API
requests==2.32.3

# Development & Testing
pytest==8.3.4
pytest-django==4.9.0
